# swagger/doc.json
{
  "schemes": ["http"],
  "swagger": "2.0",
  "info": {
    "description": "Backend API for RazorBlog including Authors, Blogs, Comments, Shares",
    "title": "RazorBlog API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "RazorBlog Support",
      "url": "http://razorblog.io",
      "email": "support@razorblog.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
    "/authors/login": {
      "post": {
        "description": "Logs in an author and returns JWT token",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Authors"],
        "summary": "Author login",
        "parameters": [
          {
            "description": "Login credentials (email, password)",
            "name": "credentials",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/authors/register": {
      "post": {
        "description": "Creates a new author account",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Authors"],
        "summary": "Register a new author",
        "parameters": [
          {
            "description": "Author info (name, email, phone, password)",
            "name": "author",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/authors/{id}": {
      "get": {
        "description": "Retrieves an author by their ID",
        "produces": ["application/json"],
        "tags": ["Authors"],
        "summary": "Get author by ID",
        "parameters": [
          {
            "type": "string",
            "description": "Author ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      },
      "put": {
        "description": "Updates an author's info",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Authors"],
        "summary": "Update an author",
        "parameters": [
          {
            "type": "string",
            "description": "Author ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Updated fields (name, email, phone, password)",
            "name": "author",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      },
      "delete": {
        "description": "Deletes an author by ID",
        "produces": ["application/json"],
        "tags": ["Authors"],
        "summary": "Delete an author",
        "parameters": [
          {
            "type": "string",
            "description": "Author ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
  },
...
}